% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiftool_flatten.R
\name{flatten_subject}
\alias{flatten_subject}
\title{Flatten hierarchical keywords}
\usage{
flatten_subject(
  paths,
  csv_execute = TRUE,
  csv_path = "~/Pictures/subjects.csv",
  delete_original = FALSE,
  recursive = FALSE
)
}
\arguments{
\item{paths}{List of photos to be modified. Needs a character vector with full file names or a directory name. In the latter case, you need to set \code{recursive} accordingly.}

\item{csv_execute}{logical, to determine whether modification should be directly written or returned as tibble}

\item{csv_path}{path and file name of the output csv file}

\item{delete_original}{whether the original copies of the photos should be deleted afterwards or not}

\item{recursive}{tells exiftool whether to read files recursively or not. Choose \code{TRUE} if you use a directory path instead of a list of files for the \code{paths} parameter.}
}
\value{
depending on param csv_execute:
\itemize{
\item if TRUE,  writes the tag values as csv file and writes them via exiftool to pictures
\item if FALSE, returns the tags as tibble
}
}
\description{
Write flat keywords and subject tags derived from hierarchical ones (tag XMP:HierarchicalSubject)
into photo metadata. Only writes the lowest level keywords back to the pictures. Hierarchical keyword tag is not modified.
Also deals with rating: aligns EXIF and XMP values and sets an additional keyword for Apples Photo App;
in order to deal with the non-existing ratings there - but only for export photos.
(There is no metadata flag for Apples favourites -> must be set manually based on this keyword).
It uses per default the csv option of exifool to handle different values for different files,
which is much faster than a for loop which calls exiftool every time.
Exiftool per default creates copies of the original files (*_original); this behaviour can be modified by parameter.
}
